Fix bugs found in full-codebase parallel review
Production codec (WirehairCodec.cpp), all behavior-neutral (verified by
deterministic A/B decode comparison and ASan/UBSan fuzz across N x bb):
- FreeMatrix() now nulls the derived GE pointers (_ge_matrix, _heavy_matrix,
  _pivots, _ge_row_map, _ge_col_map) so ResumeSolveMatrix()'s prior-state
  guard fails closed after a regrow allocation failure instead of writing
  through dangling pointers.
- BackSubstituteAboveDiagonal window_row_limit no-limit sentinel raised from
  0x7fff to 0x10000: row indices are uint16_t and pathological defer counts
  can exceed 0x7fff, which would silently skip light rows and corrupt output.
- AllocateWorkspace/AllocateMatrix/AllocateInput reject sizes that truncate
  in a 32-bit size_t (OverrideSeeds with large dense counts could pass the
  ChooseMatrix guard but overflow the workspace size).
- Peel structures are now placed at an 8-aligned offset after the recovery
  blocks: odd block_bytes previously misaligned every PeelRow/PeelColumn/
  PeelRefs access (UB caught by UBSan; benign on x86 only).
- CAT_HEAVY_WIN_MULT window update in Triangle() uses memcpy instead of a
  reinterpret_cast'd uint32_t* that is misaligned whenever
  first_heavy_column % 4 != 0 (same codegen on x86, defined everywhere).

Table generator certification bug (tables/HeavyRowGenerator.cpp):
- IsFullRank() eliminated columns i..5 while reading pivots from the offset
  submatrix (columns 12..17), so the shipped kHeavyMatrix "perfect under
  binary perturbation" 1M-trial certification never tested anything.  With
  the corrected elimination the matrix goes singular at ~0.39%/trial, the
  random GF(256) 6x6 floor (~1/256), so the claimed property is unattainable
  for ANY seed.  The generator is now a regression gate that verifies it
  reproduces the shipped wire-format bytes and reports the real perturbation
  rate; the misleading comment in WirehairCodec.cpp is corrected.  Shipped
  matrix bytes are unchanged (wire format preserved).

Bench/tools/tests:
- wirehair_v2_bench: seedtable/densecheck/densetune/densecount/densegrid/
  compare now reject --N outside [2,64000]; GetDenseCount extrapolates past
  64000 and GetDenseSeed then reads past the 100-entry kDenseSeeds table.
- test/SiameseTools.cpp: GetTimeUsec() macOS branch widened before multiply
  (32-bit product wrapped every ~4295s, breaking timings and entropy seeds).
- tables/GenerateMostDenseSeeds.cpp: ENABLE_FULL_SEARCH path referenced
  kDenseSeedCount before declaration (build break when enabled).

Experiments:
- precode_sim.cpp Rng::Below used 2^64 mod bound instead of 2^32 mod bound
  as the Lemire rejection threshold (bias was negligible, ~bound/2^32, but
  the primitive claimed exactness).
- rank_total.py crossover_ratio: schemes with lighter heavy work but heavier
  base now correctly report "beats above this ratio" instead of "never
  beats" (latent; no recorded result had that sign pattern).

Validation: unit_test deterministic warnings match baseline bit-for-bit
(122-point N=5648/6162 repro identical pre/post), large_message_test,
wirehair_v2_policy_test, ASan/UBSan decode fuzz (zero findings post-fix),
precode_sim/rank_total/peel_sweep self-tests all pass.

+++ experiments/precode/precode_sim.cpp
@@ -75,7 +75,10 @@ struct Rng
         if (bound <= 1u) {
             return 0;
         }
-        const uint32_t threshold = (uint32_t)((0u - (uint64_t)bound) % bound);
+        // 2^32 mod bound, computed in 32-bit arithmetic: 0u - bound wraps to
+        // 2^32 - bound.  (Widening to 64 bits first would compute 2^64 mod
+        // bound, which is the wrong rejection threshold for a 32-bit lane.)
+        const uint32_t threshold = (0u - bound) % bound;
         for (;;)
         {
             const uint32_t x = U32();
